refactor: move controller.common to support.controller
