+++ src/main/java/com/meoguri/linkocean/controller/bookmark/CategoryController.java
@@ -4,8 +4,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.meoguri.linkocean.controller.common.ListResponse;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
+import com.meoguri.linkocean.support.controller.ListResponse;
 
 @RequestMapping("/api/v1/categories")
 @RestController
