+++ src/main/java/com/meoguri/linkocean/controller/notification/NotificationController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.meoguri.linkocean.configuration.security.jwt.SecurityUser;
-import com.meoguri.linkocean.controller.common.SliceResponse;
 import com.meoguri.linkocean.domain.notification.entity.Notification;
 import com.meoguri.linkocean.domain.notification.service.NotificationService;
+import com.meoguri.linkocean.support.controller.SliceResponse;
 
 import lombok.RequiredArgsConstructor;
